Add files via upload
Loading the 32x32 dataset: Load the data generated in Reducing Image Sizes to 32x32.
Create Callback for F1 Score: F1-macro score is the official metric of the competition. We create a callback to keep track of that value as we train the model.
Creating and Training the Model: Create a simple model (taken from the official Keras tutorial) and train it.
Evaluation: Display the plots from the training history.
Submission: Run predictions with model.predict, and create submission csv file.